Split out Rust FFI code generation into separate files. Rename wrapper function generators for clarity.
